ENH: Add an additional test file anchors.html for more extended testing
This file should be used to test for all bad references to be detected,
which is even as of current v9.4.0-38-gc9cbc276 which I thought fixed all issues,
is not complete:  at times there is a false positive for use of an anchor ahead
of the definition of the element (i.e. here the anchor1.

TODO: figure out how to feed linkchecker this entire file instead of a single
url for checking as it was done in the test_anchor